Fix CI compile errors from 2b's first head
- SearchVector's VectorQuery parameter shadowed the package's q()
  identifier-quoting helper; renamed to vq (the body's q(table)/
  q(column) SQL-quoting calls resolve again).
- Incarnation.Version is int64 while schema versions and the legacy
  adapter's expectedVersion are int: cast at the three boundaries
  (TableState, Migrate/PlanMigration, legacyStore).
- mustNS landed inside a multi-line httptest.NewServer composite in
  local_embed_test.go; moved after the complete statement.

+++ internal/store/legacy.go
@@ -48,11 +48,11 @@ func (l legacyStore) DropTable(ctx context.Context, nsName, table string) error 
 }
 
 func (l legacyStore) PlanMigration(ctx context.Context, nsName, table string, changes []schema.Change, emb Embedder, expectedVersion int) (*MigrationPlan, error) {
-	return l.Store.PlanMigration(ctx, nsName, table, changes, emb, Incarnation{Version: expectedVersion}, nil, Incarnation{})
+	return l.Store.PlanMigration(ctx, nsName, table, changes, emb, Incarnation{Version: int64(expectedVersion)}, nil, Incarnation{})
 }
 
 func (l legacyStore) Migrate(ctx context.Context, nsName, table string, changes []schema.Change, emb Embedder, expectedVersion int) (*schema.TableSchema, error) {
-	return l.Store.Migrate(ctx, nsName, table, changes, emb, Incarnation{Version: expectedVersion})
+	return l.Store.Migrate(ctx, nsName, table, changes, emb, Incarnation{Version: int64(expectedVersion)})
 }
 
 func (l legacyStore) ListMigrations(ctx context.Context, nsName, table string) ([]Migration, error) {
